SQLite Storage Option
Implements another option for storing the data.

Currently, one of the pain points of flat files is that FTP/SFTP transfer of the files take a long time. When moving server files around, it's often that transferring the data takes up a significant portion of the time.

However, with MySQL most shared providers have limited amount of SQL databases you can create, so if existing plugins already require it, then you don't be about to use the MySQL storage. Additionally, it would be nice to have a self-contained storage like SQLite instead of having to rely on a separate database server.