Don't check for updates when turned off / on main thread
The plugin still makes the http call to check for the update even while the update checker setting is turned off. It just won't display the result.

Moved the check off the main thread or else the server will stall when there's something wrong with the http connection. If there is slow connection or the spigot website is down, the server will stall or crash.